fix: prevent git progress writer race reading stageNumber
Since the progress writer runs in a goroutine, it can call the write
function after the git operation has exited and while the next stage is
starting, resulting in a race. This fixes the race by using a second
variable and closing the writer early.

An alternative fix is to rewrite the progress writer to be synchronous.